Fix a bug that caused important debugging info to be deleted in modular solver.
The question is whether or not cabal-install will tell you why it tries
a particular goal (i.e., because it's the dependency of something else).

This regression was introduced between 0.14.0 and 1.16.0 of
cabal-install in order to fix a space leak. Restoring the information
actually costs some space again (although not as much as was required
back in 0.14.0).  I think the extra info is worth it though.