Don't make cabal-install into a library just yet
At some point in the future we probably do want a library or two
but we should make that decision deliberately and think about the API
that we want to expose.

I understand the change was made not to expose a lib API but to avoid
compiling the modules twice for the test-suite. It is true that that's
annoying but we don't yet have support for private or "convenience"
libraries and installing it as an exposed lib is not what we want to do
at the moment.