More diff friendly pretty printing of cabal files
Added the new field constructor functions 'listFieldWithSep'
and 'commaListFieldWithSep' to be able to define pretty printing
for fields, where each value of the field is positioned on a new line.

Which fields are nested is currently hard coded in the 'ppField' function.

(cherry picked from commit 5772503b33d945ae695e9b83730eedeed5a66372)