Treat all flags of a package as interdependent.
There was an open bug that could cause conflict sets to be
computed incorrectly if flags were involved in the conflict
set. Incorrectly computed conflict sets can lead to the solver not
finding solutions even though they actually exist.

Because I decided that computing flag dependencies correctly is tricky
and probably not worth the effort, we're now treating all flags (of a
package) as mutually dependent. This means that if any flag choice ends
up in a conflict set, then it's as if all flag choices (of the same
package) are in that conflict set. This is potentially slower for
packages having many flags, but it should at least be correct, because
it's conservative.